Refactor itrc command
